Add govuk-delivery and worker to Pinfile and Procfile
I've copied these commands from the Procfile in the app repo. I haven't made
the worker a dependency of the app because the app startup script creates the
venv, without which the worker won't run.

I've gone with hyphenated names for these to match the app name as it's
deployed, rather than the repo.