Add MySQL on ramdisk
We had a performance issue with running tests with MySQL in that they
were taking a substantially long time to run, and it seems to be pegged
by the disk.

To stop blocking developers do their work, and because we don't care
about the MySQL data in CI, a solution was to devised to run MySQL in
ramdisk.

This Puppet requires an initial bootstrap in that we'll have to remove
the MySQL data, and then run Puppet, but then after that it'll mount the
directory on tmpfs, and then when Puppet runs the ::mysql module will
create the data structure.

The directory also needs to be owned by a MySQL user, and on first boot
that user may not exist, so I've added a user creation for the 'mysql'
user. This does not seem to cause any issues with the main mysql puppet
module.

On boot I have added a cronjob to ensure that Puppet runs immediately so
that there is a working MySQL directory ASAP.