Remove hstore extension from email-alert-api's db definition
We've finished migrating the app to use json columns instead of hstore to
store the main matching criteria for email subscription lists.

This change isn't enough to actually disable the extension in the database (we
don't support ensure => absent in our puppet module for this) so we'll devops
it in each environment (`DROP EXTENSION hstore;`).

We've done similar things before: 3b8fc618febac824edf24aae719d5c2709ce8434