Replicate mirrors to S3
We want to serve mirrors from S3 rather than from VMs.

This commit adds S3 as a target for the cron job that copies to our
mirrors. To achieve this the AWS secret & access keys are set by
`govuk_setenv` and the copy itself done using the AWS command line
tool.