Use shallow clone of deployment repo
Cloning the full history of the deployment repo takes a long time due
to the large changes to GPGed files. We only need master, so only clone
that.

Shallow Clone doesn’t appear to be compatible with wipe-workspace as it
forces a `git fetch` without specifying a `depth`, so replace it with
cleaning the workspace before checkout.