Remove rake tasks to run migrations
Now that the content-performance-manager is being deployed to production, we can
rely on the over night "copy data" jobs to populate data in staging and
integration.

Also, because a "blank" database will no longer be copied from production, we no
longer need to run the migrations in these environments every morning.