Add exec to init MySQL on initial mount
The MySQL puppet module requires MySQL to be running before any changes
can be made. MySQL will not run until the datadir has been correctly
initiated, which it won't be if the directory has just been mounted on a
ramdisk.

This creates an exec which runs mysql_install_db which initiates that
directory, but only after it's been mounted and that the package exists.
The mount notifies the exec itself, and if the `mysql` database already
exists, it won't run (as it assumes it's already been initiated with the
meta tables).