Reload configuration rather than restart
In an attempt to stop interupting jobs a class was created that issued a
"safe-restart" from the jenkins-cli. Unfortunately this does not appear
to work with pipeline jobs, which is problematic on our CI environment.

This changes the command to use "reload-configuration" which reloads all
configuration that may have been made to disk, and should *hopefully*
not interupt any running jobs when it does it.