Use SSH credentials to access Github from CI
During busy parts of the day we have been hitting the Github rate
limit of 5,000 requests/hour. This causes an exception in Jenkins
and stops jobs building.

This commit changes the Multibranch Pipeline configuration to use
separate SSH credentials for checkout as to scanning. This should
mean that we are less likely to hit the rate limit.