Add parameter for max_connections
PostgreSQL has a maximum number of connections that are available for
applications to connect to. The default maximum is 100, but for a busy
production database this may be too low.

If applications also do not properly close off their connections (and
therefore persist in an 'idle' state) then this can fill up the amount
of available connections.

If there are no more connections available, then it may block
applications from connecting to the database.

This adds an alert to let us know when we're close to the threshold. In
this instance, we should check the number using `psql`:

`SELECT state, count(*) FROM pg_stat_activity GROUP BY state;`

We could then terminate some of the older connections.