Add templating of Google credentials file
Google Cloud does things differently in that you must have a credentials
file to be able to authenticate with their API. This causes a problem as
while we can provide a file to use for authentication, we want to ensure
that when people run the Jenkins job they are passing in some form of
authentication, because otherwise they could potentially cause serious
impact in deployment of DNS.

This looks to template this credentials file, in which we can store most
of the secrets of the file in encrypted hiera, but also prompt for a
couple of parts of the file in the Jenkins job itself, and pass these in
as environment variables. These secrets can be stored in our password
store to use.