Restart carbon-aggregator service every 5 minutes in Staging
Carbon-aggregator connections in Staging seem to be filled up from clients on
the licensify VPN, and the errors in the log file are filling the disk.

This is probably caused by network problems that are being investigated
with Carrenza, for now we can configure this cron job to restart the daemon
every 5 minutes.