Only reload configuration on Jenkins 2
To use the reload-configuration CLI command you require "credentials" to
be added in the Jenkins GUI. In Jenkins 2, you're able to create a set
of credentials that has an SSH key for "system scope", but in Jenkins 1
this option isn't available.

This commit adds a conditional that only allows reload on Jenkins 2. The
other classes that use the govuk_jenkins::reload class are only Jenkins
2, so this just fixes Jenkins 1 machines (which generally weren't
affected by service restarts as badly as our CI machines).