Serve an uncached empty gif for analytics purposes
Configure `assets-origin.domain/static/a.gif` to always serve an empty
gif that isn’t allowed to be cached. This is to be used in conjunction
with a new GOV.UK analytics tracker. Javascript will insert an image
tag on every page referencing this, with tracking data added to the
query string of the gif. This will be picked up in CDN logs.

The anti-caching might be slightly too aggressive here, but the
overhead is minimal and debugging this in production would be hard.
It’s easier to scale back the aggression once it’s live.