Update bucket to destination
When enabling backups for S3 I found that the backup process was being
inconsistent:

 - backups were being taken as full backups each time
 - backups were not being tidied up and were leaving orphans and
 incomplete sets behind

When trying to run `duplicity collection-status` against these backups
it failed because it didn't work with the URL that was defined in the
original job:

s3+http://govuk-offsite-backups-production/

Changing this to use the fully defined bucket name allowed me to cleanup
and check the status of a backup. In the actual job, just having the
defined "s3://foo" still successfully took a backup as the boto backend
was still used.