Add virus scan speed alert and Grafana dashboard
Add an Icinga check to alert when the waiting time in the asset-master queue
to scan documents is longer than 20 minutes (warn) and 40 minutes (critical).
We might need to adjust these values when we start analyzing the new metrics.

Add initial Grafana dashboard to visualise number of documents and time since
the document was uploaded until it's ready to be scanned.