Don't lint release branches
We lint changes by taking the diff with master. This works for dev branches but
not release branches, which are either behind or in sync with master.

When the commits on the current branch are already in master, diffing it to
master makes no sense.  govuk-lint-ruby picks up lines that wouldn't have been
linted when the changes in it were first made.  This is happening for the
deployed-to-production branches of our apps, and leads to a lot of spurious
lint errors when changing the content schemas, because we trigger CI jobs using
production app code with the new schemas.

This is the same situation as described in https://github.com/alphagov/govuk-puppet/pull/5396