Access Jenkins build parameters with `params.FOO` instead of `env.FOO`
There are three ways of accessing build parameters in Jenkins pipelines:
`env.FOO`, `FOO` and `params.FOO`. But the first two versions don't work
on the very first build of a branch (see JENKINS-40574).

We have been using `env.FOO` and working around the problem by calling
the `initializeParameters` function. But it is simpler to use
`params.FOO` instead. We should be able to delete the
`initializeParameters` function once we've updated every other project
to using `params` rather than `env` for build parameters.