Evaluate boolean build parameter correctly in Jenkins lib
We recently switched from `env.IS_SCHEMA_TEST` to
`params.IS_SCHEMA_TEST` to check whether a build is being run to check
changes to the content schemas.

The problem is that these variables have different types: the `env`
version is a string, and the `params` version is a boolean. This means
that `params.IS_SCHEMA_TEST == "true"` always evaluated to `false`,
which means that schema tests were not being run correctly.