Push version tag last in publishGem
I think this works better when parts of the publishGem proecess fail.

Skipping publishing the gem if the version tag exists prevents
retrying gem publication if it failed the first time. This is because
the second time the job runs, publishing the gem will not be attempted
as the tag already exists.

Additionally, if the gem publication fails, the tag being pushed is
misleading.