In publishGem, check if the tag exists on origin
This is what matters, as the purpose of this code is to push a tag, so
check if the tag exists on the remote, rather than locally.

This helps to prevent issues where the tag might exist locally, and
therefore the code is skipped, but it doesn't exist on the remote.