Make pushing a release branch work, even if the branch doesn't exist
Even though this is being done in the pushTag function, it may as well
work for new repositories that may not have a release branch for the
first release.

I think this change may be only necessary with the version of git in
use on CI (1.9.1) as I cannot reproduce the issue locally with
git (2.11.1).