Don't return if the gem is published or the version tag is present
This means that rerunning a failed job will fix a missing tag if
publishing the tag failed. As the tagging is now after the gem
publish, it doesn't have any effect, but removing it is more
consistent, and will matter if other actions are added after pushing
the tag in the future.