Jenkins Deploy Upgrade
Upgrading deploy Jenkins to version 2.

Plugin versions had to be modified to be compatible with upgraded Jenkins core.

plugin=github-oauth must be omitted when provisioning a vagrant instance.  This is because the plugin will cause Jenkins to not load its configuration due to the lack of an auth backend.

See: https://trello.com/c/RZCVbiHd