Don't supress postgres errors
When importing a postgres database we sent all errors from postgres to /dev/null
This is not very helpful. Errors in this step should be logged and cause replication to fail.

As a first step I've made it log to a temporary file instead.

The "role doesn't exist" errors are safe to ignore but I had trouble getting it
to pipe properly and it seemed reasonable to include in the log file.

Also, clean up old elasticsearch indexes after importing elasticsearch data.
These take up a lot of disk space that can cause the thing to fail.