Update crawler S3 sync command to avoid re-uploads
When syncing the website mirror files to S3, s3cmd is not detecting the
remote files so it's re-uploading all the files every time it runs.

s3cmd is quite specific about the declaration of the origin and target
directories, this change removes the dot at the end of the origin
directory so the files in the remote dir can be detected. With this change
we also need to force the target directory to finish with /

More info:
https://github.com/s3tools/s3cmd/issues/745
https://github.com/s3tools/s3cmd/issues/660