Filter coverage output to just a summary
The rspec-puppet coverage module outputs a list of untouched
resources after it's produced the summary of the coverage. Given the
number of modules with limited coverage this produces a huge amount
of output. This filters the coverage output to just the summary unless
the `FULL_COVERAGE_REPORT` environment variable is set.