Update default `branches_to_exclude` from CI
We were excluding `deployed-to-*` branches from being built in CI by
default. This included `deployed-to-production` which we are using for
schema testing now and this was consequently being overriden by almost
all of the `deployable_apps` in `hieradata/common.yaml`.

This commit updates the default to be more specific and removes the
repetitive overrides.