Remove absent rbenv version
When running tests locally on OSX we found that we kept running into
this puppet-rspec failure:

```
Failure/Error: it { is_expected.to compile }
       error during compilation: Parameter ensure failed on Package[rbenv-ruby-2.1.2]: Provider must have features 'purgeable' to set 'ensure' to 'purged'
```

This seems to be because OSX package provider is not able to delete
packages and it causes problems with the upstream module, although there
are questions because we explicitly set the ::operatingsystem and
::osfamily facts in
`spec_helper.rb` and we don't have issues with other packages.

This change should now be deployed, so we should entirely remove it.