Temporarily add a rake task to Integration data sync complete job
Until this rake task has been run in Production, we want it to run on
Integration each day before the email-alert-api sync_govdelivery_topic_mappings
one. This will save us several hours of deleting and recreating the same data
in GovDelivery each day.

I've put the CSV file at this path on both whitehall backends in Integration,
and will delete them once we're finished with this.