Only try to ensure permissions if present
If a user leaves then we can set their user to "absent" in the
govuk_user defined type.

This will remove the user, but we don't want to delete their home
directory for $reasons. This will only try to set permissions on their
home directory if a user is present, but will still remove their SSH key
directory.

Due to the conditional the ordering statement had to be rejigged.