Add more flexibility to shared Jenkins `buildProject` task
Add three parameters to `buildProject` in the jenkinslib. These are all
parameters which contain code to be executed at different stages of the
build:

 - `beforeTest`, which is run immediately before the tests
 - `overrideTestTask`, which is run in place of the default test task,
   `bundle exec rake`
 - `afterTest`, which is run immediately after the tests

This will allow us to use `buildProject` in projects which have builds
with extra steps, such as Jasmine tests that cannot be added to the
default rake task or Pact report publishing.