Convert Jenkins `buildProject` parameters to map of named options
In preparation for adding more parameters, replace the current parameter
list with a groovy Map.

This will allow individual Jenkinsfiles to just specify the parameters
they require, and it will be more readable because the parameters will
be named.

Continue to handle the original `sassLint` parameter, at least until all
the Jenkinsfiles that use it have been updated to use the options
parameter.

No Jenkinsfiles are using the `extraRubyVersions` parameter yet, so it's
safe to just merge it into the map.