Run govuk-delivery script to update data after the data sync
This script was added in https://github.gds/gds/govuk_delivery/pull/29. Having
the data set up consistently in all environments will help us be confident
that the email-alert-api matching is working correctly before we switch
Whitehall over to use it.

I'm adding it to the data_sync_complete jobs rather than as a separate job to
be triggered by them, because it looks like triggering a build without
parameters from trigger-parameterized-builds doesn't work. This is a temporary
addition anyway - when we retire govuk-delivery this will disappear too.