Remove now unused MONGODB_NODES env var for manuals-publisher
We switched manuals-publisher to use `MONGODB_URI` in this PR [1].
This change was deployed to production at approx 11:50 on 31 Mar 2017.
So it should now be safe to remove the `MONGODB_NODES` env var.

In order to ensure I didn't miss anything, I created this commit by
reverting the following two commits and squashing into one:

* "Write MONGODB_NODES for manuals-publisher correctly" [1]
* "Provide a list of mongodb hosts for manuals-publisher" [2]

[1]: alphagov/manuals-publisher#936
[2]: 06acc5ce08e11bee1d05d11e47c6afacb073a7f0
[3]: 0d03bfd5c6e5e875906c9b1a9f44969f6dd2c330