Run fake virus scan in development VM
This commit adds a cron job for the development VM to run a fake virus scan every minute. The fake scan just copies files from the `upload` folder to the `clean` folder to simulate a virus scan.

This allows users of the development VM to do things like publishing documents with attachments in whitehall without having to run the rake task manually each time.