Update app SSL certificates for dev VM
The SSL certificate and key in `hieradata/development_credentials.yaml`
is added for each app in the dev VM and enabled in the nginx config via
a flag in `hieradata/development.yml`.

The existing certificate had been generated for `*.dev.alphagov.co.uk`
and consequently could not be used in the dev VM which no longer uses
that domain.

This commit updates the certificate with the correct domain
`*.dev.gov.uk` which will enable us to use `https` as part of work to
allow apps to be testing in `Rails.env == 'production'` in the VM.

Other exciting facts about the new certificate:

```
Certificate:
    Data:
        Version: 3 (0x2)
        Serial Number: 17279162745291281828 (0xefcbeaf7c94015a4)
    Signature Algorithm: sha256WithRSAEncryption
        Issuer: C=GB, ST=London, O=GOV.UK Development (self-signed),
CN=*.dev.gov.uk
        Validity
            Not Before: Apr 10 10:26:17 2017 GMT
            Not After : Apr  8 10:26:17 2027 GMT
        Subject: C=GB, ST=London, O=GOV.UK Development (self-signed),
CN=*.dev.gov.uk
        Subject Public Key Info:
            Public Key Algorithm: rsaEncryption
                Public-Key: (4096 bit)
```