Add Redis config for Frontend
The new emergency banner implementation uses Redis. This exposes the env
vars needed for Frontend to access Redis. Although most of the code to
display the emergency banner is being moved into Static, Frontend still
needs to selectively display the promo banner if there is no emergency
banner to display. Checking for the keys in Redis will allow Frontend to
achieve this.