Allow disabling of database rename on replication
This commit allows passing the `-o` argument to the development VM replication script to prevent the database names being renamed from `*_production` to `*_development`.

This can be used when running apps with `RAILS_ENV=production` so that the apps find their databases.

This will be used when setting up the training environment since that will be running as a production environment.