Add building of Docker images in Jenkins
If a repository has a Dockerfile, build it. If tests pass (in the
traditional way, not using Docker yet), push the image to the public
Docker hub tagged with the branch name. If it’s the master branch, also
push an image tagged with the same name as the release tag, e.g.
`release_123`.