Merge and update development-vm documentation
This was in a bit of a state since merging the development repo into
govuk-puppet. There were two seperate READMEs in this folder, with some
significant overlap between them.

Here we have attempted to merge the two guides into a document with a
straightforward purpose: getting started as a new gov.uk developer.

It covers installation of Vagrant and bootstrapping a VM, creating and
sharing your credentials across the various services, and getting the
gov.uk application stack up and running.

As often happens after taking something apart and putting it back
together, there are some left over parts.  There were a fair few
troubleshooting tips and "what-if" scenarios covered in the original
documentation that I have collated into a page in the Dev Manual
to be reviewed seperately.