Create CLI class
This commit creates a CLI class that can be called by a user when they
want to interact with Jenkins using jenkins-cli.jar. It was already
invoked by the reload-configuration class, so I have moved the pre-reqs
into it's own class and created a very basic script which can be
invoked.