Add draft-router and draft-router-api
This commit adds draft versions of router and router-api to the development VM. It follows the same pattern as other draft apps (e.g. government-frontend). This completes the draft stack in the development VM. It also updates the replication script to replace backend routes for draft-router with dev.gov.uk ones.