Delete mostly-empty and broken deployment dashboards
Delete dashboards which are missing most of the data or are not
reporting it correctly. This errs on the side of deleting rather than
keeping dashboards to avoid confusing developers with missing data while
the dashboards are still so new. Any dashboards that we fix can easily
be re-added later.

 - `authenticating-proxy` does not seem to report 5xx responses or error
   data in the same way as other Rails apps
 - `contacts` conflates admin and frontend data, and does not report
   Errbit errors correctly
 - `govuk_content_api` is deprecated
 - `govuk_need_api` reports graphite data as `need_api`, so it does not
   appear on the dashboard, which expects the data to match the app name
 - `policy-publisher` is not reporting 5xx data correctly