Clean-up the Pinfile and Procfile for development
This commit cleans-up the Pinfile and Procfile that have been worked on recently.

It sorts the Pinfile in alphabetical order, ensures consistency of markup, removes the archived `dfid-transition-worker` and adds a pseudo-process named `draft-origin` that starts `draft-router`, to go along with the pseudo `www` process.

It ensures consistency of commenting and formatting in the Procfile and removes the archived `dfid-transition-worker`.