Hide error/slow responses graphs if data is missing
The graphs at the bottom of the deployment dashboard come from
elasticsearch. For a lot of dashboards, these are empty because the
events aren't logged in a consistent way.

For apps where the data is not available, hide these panels. Also hide
them for low usage apps where there won't be enough events to see
trends.

In the case of the error graph, the graph itself is only useful when
there are errors and you want to break it down by controller, so we
might want to change how we present this for apps that don't error all
the time. A series is only drawn on the graph if there is an event so
the graph is just blank if everything is fine, which is confusing.

https://trello.com/c/lVNtP9A4/47-hide-errors-by-controller-and-slow-responses-by-controller-graphs-on-dashboards-where-they-have-no-data