Update govuk_docker class
Update the govuk_docker class to enable some options that we use
elsewhere in the code, so that the stuff we do for docker in CI is able
to use the govuk_docker wrapper class and get monitoring and packages
contained in that class.

Also added a tag to always install the latest version of docker.