Add content training environment
This commit adds all the configuration required to build a content training environment based on the development VM. The environment can be built using either Vagrant or Terraform in AWS. There are a number of provisioning scripts that are run as part of the build. A Jenkins job is also defined to plan/apply/destroy the VM as required, and this is scheduled to run once a week.