First draft at managing apps as containers
This commit adds a defined type to pull and run a specified container.

It's function is meant to be the beginning of a construction similar to
the govuk::app class, but does the minimum neccessary to pull an image
with the given tag, and run that image.

It allows adding environment variables using encrypted hieradata. These
will be placed in an init upstart script for the time-being before we
use proper secret management in docker.

We're using the release app for testing our processes.