Add `fast_finish` to our travis options
From https://blog.travis-ci.com/2013-11-27-fast-finishing-builds/

"Travis CI will mark your build as finished as soon as one of two
conditions are met: The only remaining jobs are allowed to fail, or a
job has already failed. "

This will speed our builds up in a number of cases as we have multiple
options that are allowed to fail.