Simplify `buildProject` `options` parameter in Jenkinslib
Restrict the `options` parameter to always be a Map of options. This
used to accept either a Map of options or a single Boolean parameter
which toggled Sass linting, to preserve backwards-compatibility with
projects which called `buildProject(false)`.

All projects which used the Boolean parameter have been updated to use
the Map, so it's safe to remove the Boolean usage.