Add Release app envvars for containers
This adds the environment variables for the Release app running in a
container. I am aware that the DATABASE_URL envvar isn't a great way of
defining the variable and that we could do it in a better way, but in
the interests of proving how the apps run in containers this should be
sufficient.