Add jenkins jobs to run search benchmarks
These tests run the rummager healthcheck, which benchmarks search
against a set of judgements. One job checks for the presence/absense of
particular documents in query results, and the other tests that the "did
you mean" returns suitable suggestions for misspelled queries.

These used to run on the old CI server, which we replaced. These take a
long time to run, so we're using the deploy jenkins and running them
nightly instead of incorporating it into the rummager CI pipeline.