Change the way image tags work
The Puppet code that used the ::docker::image and ::docker::run classes
did not fit into our current deployment model as it meant Puppet
managed which images were running, whereas we need a job which deploys a
newer image and then restarts the service.

the ::docker::run defined type is useful as it creates an init script,
so we're going to have a generic "current" release tag, and then use `docker
image tag` to point this tag to a specific image.