Add logspout docker container
Logspout is a docker container for shipping logs. This adds the
dockerfile so it can be used as such.

Logspout has been added to the govuk_docker module, rather than the
govuk_container module, as it will be a hard depenency of govuk_docker.
Whenever we use docker we want to have logging set up for it.