Make estimated disk space for the VM more realistic
Once you build the VM and replicate all the data you're eating closer
to 70GB for the VM and 30GB for the replication backup.  You can delete
the backup to reclaim that 30GB, but you still need to have that much
spare so that you can run another replication later when things get
out of date.

The VM is configured to grow up to 100GB so it's possible that you'll
actually need closer to 130GB free space on your host drive, but 100GB
seems more realistic.