Fix replication logic
The $DRY_RUN, $SKIP_MONGO and $SKIP_ELASTIC arguments of the
dev vm replication script were not having the expected effect due to an
error in the logic.

This commit fixes the logic to ensure that the mongo munging and
elasticsearch index deletion are not run when they should be skipped or
during a dry run.