Rate limit contact/govuk urls for POST requests only
We were rate limiting /contact/govuk/service-feedback and
/contact/govuk/problem_reports already, but for all requests.  The problem
is that /contact/govuk and /contact/govuk/foi are both restful and so we
want to allow the GET requests through, but limit the POST requests that
actually submit.  To do this we introduce a new map that toggles on both
the Rate-Limit-Token header and request method.  This map sets up a
variable that is blank for anything that isn't a POST request, and also
blank for a POST request that has a correct Rate-Limit-Token header.  All
other requests have the binary version of the remote address as the value.

This variable is then used in the "contact" `limit_req_zone` to decide which
requests to rate limit.  It'll let all blank values through, but rate limit
any others to 6 requests per minute (with a 5Mb cache).  Note that the limit
zone is unchanged, apart from the variable used as the key.

We then add all POST requests for the feedback app to the existing location
that uses this `limit_req_zone`.  We'll need to keep this in sync with any
new POST endpoints in the feedback app.