Add calendars as dependency to publisher
The default message for the fact-check email in publisher needs to know how to
calculate working days and it does so by reading one of the bank holidays json
files hosted by the calendars app.  This seems like a small dependency, but the
default message for the fact-check email is rendered on every visit to the
email edit form.  If calendars isn't running then we break publisher which
would probably be painful in dev.