Allow pushing images to GCR
This adds a function that pushes docker images to GCR. It uses shell
commands rather than the docker plugin as I could find no way to
integrate the docker plugin with the GCR credentials file. There is a
plugin but it does not seem to support use in pipeline builds.