Fix virus scan for asset manager tests
`clamav` was added in to CI in
https://github.com/alphagov/govuk-puppet/pull/5566 to make the asset
manager tests run. This commit adds a piece of configuration that will
make `govuk_clamscan` point at the non-daemonised version of clamscan.
The daemon version is run by the `clamav` user and won't be able to
scan the files uploaded by asset manager's integration tests, causing
test failures.