Add safety flag for transition logs
I needed a way to safely test the function of transition logs without
causing any production issues. This commit adds an "enable_cron" flag
which when enabled starts the job running. Without this flag there will
be no cron job and the job won't push to master. This enables us to have
Puppet create everything required in our testing environment, but
without a risk of pushing anything to the upstream repository.